Faner for simulering og brev skal låses dersom vedtakk ikke er lagret enda
