+++ src/frontend/Komponenter/Behandling/Fanemeny/Fanemeny.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react';
 import { FC } from 'react';
 import styled from 'styled-components';
-import { NavLink, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { IBehandlingParams } from '../../../App/typer/routing';
-import { filtrerSiderEtterBehandlingstype, sider } from './sider';
-import { Normaltekst } from 'nav-frontend-typografi';
+import { filtrerSiderEtterBehandlingstype, ISide, SideNavn, sider } from './sider';
 import { useBehandling } from '../../../App/context/BehandlingContext';
 import DataViewer from '../../../Felles/DataViewer/DataViewer';
 import { Sticky } from '../../../Felles/Visningskomponenter/Sticky';
 import navFarger from 'nav-frontend-core';
-import { useApp } from '../../../App/context/AppContext';
+import { Steg } from '../Høyremeny/Steg';
+import Fane from './Fane';
 
 const StickyMedBoxShadow = styled(Sticky)`
     box-shadow: 0 2px 5px rgba(0, 0, 0, 0.4);
@@ -25,37 +25,15 @@ const StyledFanemeny = styled.div`
     background-color: ${navFarger.white};
 `;
 
-const StyledNavLink = styled(NavLink)`
-    border-bottom: 5px solid white;
-    color: inherit;
-    text-align: center;
-    text-decoration: none;
-    width: 100%;
-    padding-top: 1rem;
-    padding-bottom: 1rem;
-
-    :hover {
-        border-bottom: 5px solid ${navFarger.navBlaLighten20};
-
-        .typo-normal {
-            color: ${navFarger.navBla};
-        }
-    }
-
-    &.aktiv {
-        background-color: ${navFarger.navLysGra};
-        border-bottom: 5px solid ${navFarger.navBla};
-
-        .typo-normal {
-            font-weight: bold;
-        }
-    }
-`;
-
 const Fanemeny: FC = () => {
     const { behandlingId } = useParams<IBehandlingParams>();
     const { behandling } = useBehandling();
-    const { gåTilUrl } = useApp();
+    const låsendeSteg = [Steg.VILKÅR, Steg.BEREGNE_YTELSE];
+    const fanerSomKanLåses = [SideNavn.SIMULERING, SideNavn.BREV];
+
+    const faneErLåst = (side: ISide, steg: Steg): boolean => {
+        return fanerSomKanLåses.includes(side.navn as SideNavn) && låsendeSteg.includes(steg);
+    };
 
     return (
         <DataViewer response={{ behandling }}>
@@ -65,19 +43,13 @@ const Fanemeny: FC = () => {
                         <StyledFanemeny>
                             {filtrerSiderEtterBehandlingstype(sider, behandling).map(
                                 (side, index) => (
-                                    <StyledNavLink
-                                        key={side.navn}
-                                        to={`/behandling/${behandlingId}/${side.href}`}
-                                        activeClassName="aktiv"
-                                        onClick={(e) => {
-                                            e.preventDefault();
-                                            gåTilUrl(`/behandling/${behandlingId}/${side.href}`);
-                                        }}
-                                    >
-                                        <Normaltekst>
-                                            {index + 1}. {side.navn}
-                                        </Normaltekst>
-                                    </StyledNavLink>
+                                    <Fane
+                                        side={side}
+                                        behandlingId={behandlingId}
+                                        index={index}
+                                        deaktivert={faneErLåst(side, behandling.steg)}
+                                        key={index}
+                                    />
                                 )
                             )}
                         </StyledFanemeny>
